pyproject.toml
[project]
name = "mcp-server-oracle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.5.0",
    "oracledb>=3.0.0",
    "python-dotenv>=1.0.1",
]
